Added backend user creation, login, and authentication.

+++ server/src/main/java/consamables/api/Group.java
@@ -5,35 +5,35 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Group {
     @JsonProperty
-    private Integer groupId;
+    private Long groupId;
 
     @JsonProperty
-    private Integer restaurantId;
-    
+    private Long restaurantId;
+
     @JsonProperty
     private String type;
-    
+
     @JsonProperty
     private String phase;
-    
+
     @JsonProperty
     private Integer minPeople;
-    
+
     @JsonProperty
     private Integer durationMinutes;
 
     @JsonProperty
     private Timestamp timeCreated;
-    
+
     @JsonProperty
     private Timestamp timeStarted;
-    
+
     @JsonProperty
     private Timestamp timeOrdered;
-    
+
     public Group() { }
 
-    public Group(Integer groupId, Integer restaurantId, String type,
+    public Group(Long groupId, Long restaurantId, String type,
                  String phase, Integer minPeople, Integer durationMinutes,
                  Timestamp timeCreated, Timestamp timeStarted, Timestamp timeOrdered) {
         this.groupId = groupId;
@@ -47,19 +47,19 @@ public class Group {
         this.timeOrdered = timeOrdered;
     }
 
-    public Integer getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(Integer groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
 
-    public Integer getRestaurantId() {
+    public Long getRestaurantId() {
         return restaurantId;
     }
 
-    public void setRestaurantId(Integer restaurantId) {
+    public void setRestaurantId(Long restaurantId) {
         this.restaurantId = restaurantId;
     }
