+++ server/src/main/java/consamables/jdbi/mappers/MenuSectionMapper.java
@@ -11,6 +11,6 @@ import consamables.api.MenuSection;
 public class MenuSectionMapper implements ResultSetMapper<MenuSection> {
 
     public MenuSection map(int index, ResultSet r, StatementContext ctx) throws SQLException {
-        return new MenuSection(r.getInt("menu_section_id"), r.getString("name"));
+        return new MenuSection(r.getLong("menu_section_id"), r.getString("name"));
     }
 }
