+++ server/src/main/java/consamables/jdbi/mappers/RestaurantMapper.java
@@ -15,7 +15,7 @@ public class RestaurantMapper implements ResultSetMapper<Restaurant> {
     public Restaurant map(int index, ResultSet r, StatementContext ctx) throws SQLException {
         ObjectMapper mapper = new ObjectMapper();
         try {
-            return new Restaurant(r.getInt("restaurant_id"), r.getString("name"),
+            return new Restaurant(r.getLong("restaurant_id"), r.getString("name"),
                                   mapper.readTree(r.getString("location")),
                                   mapper.readTree(r.getString("hours")),
                                   r.getString("url"));
