+++ server/src/main/java/consamables/jdbi/mappers/VoteMapper.java
@@ -9,7 +9,7 @@ import consamables.api.Vote;
 public class VoteMapper implements ResultSetMapper<Vote> {
 
     public Vote map(int index, ResultSet r, StatementContext ctx) throws SQLException {
-        return new Vote(r.getInt("user_id"), r.getInt("group_id"),
+        return new Vote(r.getLong("user_id"), r.getLong("group_id"),
                         r.getInt("minutes_interested"), r.getBoolean("can_drive"),
                         r.getTimestamp("time_placed"));
     }
